NetworkMgr: Fix PocketBook losing net access
On the PocketBook platform, for some reason, doing the `isOnline` check
(`socket.dns.toip("dns.msftncsi.com")`) without having any network
connection (`!isConnected`) results in the `isOnline` check never
succeeding again even if connectivity is later acquired. Since
`socket.dns.toip` is a simple wrapper around gethostbyname(3), we
suspect the problem is somewhere deeper in PocketBook's libc, and thus
we need to work around it somehow in KOReader.

This fix works around the problem by checking `isConnected` first before
even attempting to check `isOnline`. If the device reports not being
connected to a network, `isOnline` is (almost) guaranteed to fail
anyway. So, flip the logic in `runWhenOnline` and `willRerunWhenOnline`
to call `isConnected` first, and add `isConnected` to a few other
places.

Fixes: https://github.com/koreader/koreader/issues/10183